Moved useAzur into separate Hook
